extract gitserver main to shared package for external invocation
This is a copy-and-paste of the gitserver main function to a shared package so that gitserver can be invoked from a different entrypoint. This is how it's done for several other programs (frontend, repo-updater, etc.). There are no logic changes.
